Start refactor. Push all constants, macros, variables out to seperate files. Refactor out attract monsters screen.
