+++ attract_monsters.h
@@ -1,0 +1,16 @@
+#ifndef ATTRACT_MONSTERS_H
+#define ATTRACT_MONSTERS_H
+
+#include "constants.h"
+#include "palette.h"
+#include "wow_monsters.h"
+#include "neslib.h"
+
+
+
+/**
+ * attract_monsters() - shown while game isn't playing
+ */
+void attract_monsters(void);
+
+#endif /* ATTRACT_MONSTERS_H */
