+++ palette.h
@@ -1,0 +1,6 @@
+#ifndef PALETTE_H
+#define PALETTE_H
+
+const unsigned char palette[16]={ 0x0f,0x11,0x16,0x28,0x0f,0x16,0x11,0x28,0x0f,0x28,0x16,0x11,0x0f,0x0c,0x1c,0x2c };
+
+#endif /* PALETTE_H */
